# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EstimateRequest {
  /** Description of the construction or renovation job */
  jobDescription: string;
  /**
     * Optional location for regional pricing context (e.g., "Texas", "California")
     * @nullable
     */
  location?: string | null;
  /**
     * Contractor's hourly labor rate in USD
     * @nullable
     */
  hourlyRate?: number | null;
  /**
     * Contractor's total years of experience (can be fractional, e.g., 0.5 for 6 months)
     * @nullable
     */
  yearsExperience?: number | null;
}

export interface LaborTask {
  /** Name of the labor task */
  task: string;
  /** Estimated hours for this task */
  hours: number;
}

export interface LaborEstimate {
  /** Baseline hours for an average contractor */
  baseHours: number;
  /** Speed multiplier based on experience (e.g., 0.8 = 20% faster) */
  experienceMultiplier: number;
  /** Final estimated hours after experience adjustment */
  adjustedHours: number;
  /** Total labor cost in USD (adjustedHours * hourlyRate) */
  totalLaborCost: number;
  /** Task-by-task hour breakdown */
  breakdown: LaborTask[];
  /** Short note explaining the experience adjustment */
  experienceNote: string;
}

export interface MaterialItem {
  id: string;
  /** Material name */
  name: string;
  /** Brief description or specification */
  description: string;
  /** Material category (e.g., "Lumber", "Concrete", "Hardware") */
  category: string;
  /** Amount needed */
  quantity: number;
  /** Unit of measurement (e.g., "sq ft", "pieces", "bags", "gallons") */
  unit: string;
  /** Price per unit in USD */
  unitPrice: number;
  /** Total price for this line item in USD */
  totalPrice: number;
  /** Suggested retail store (e.g., "Home Depot", "Lowe's") */
  storeName: string;
  /**
     * Any additional notes about this material
     * @nullable
     */
  notes?: string | null;
}

export interface EstimateResult {
  /** AI summary of the job */
  jobSummary: string;
  materials: MaterialItem[];
  /** Total cost of all materials in USD */
  grandTotal: number;
  /** Labor estimate if hourly rate was provided */
  laborEstimate?: LaborEstimate | null;
  /** Pricing disclaimer */
  disclaimer: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}