# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: estimates
    description: Material cost estimate operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /estimates/generate:
    post:
      operationId: generateEstimate
      tags: [estimates]
      summary: Generate a material and labor cost estimate
      description: Uses AI to generate an itemized material list with pricing and a labor estimate based on contractor experience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateRequest"
      responses:
        "200":
          description: Generated estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    EstimateRequest:
      type: object
      properties:
        jobDescription:
          type: string
          description: Description of the construction or renovation job
        location:
          type: ["string", "null"]
          description: Optional location for regional pricing context (e.g., "Texas", "California")
        hourlyRate:
          type: ["number", "null"]
          description: Contractor's hourly labor rate in USD
        yearsExperience:
          type: ["number", "null"]
          description: Contractor's total years of experience (can be fractional, e.g., 0.5 for 6 months)
      required:
        - jobDescription
    LaborTask:
      type: object
      properties:
        task:
          type: string
          description: Name of the labor task
        hours:
          type: number
          description: Estimated hours for this task
      required:
        - task
        - hours
    LaborEstimate:
      type: object
      properties:
        baseHours:
          type: number
          description: Baseline hours for an average contractor
        experienceMultiplier:
          type: number
          description: Speed multiplier based on experience (e.g., 0.8 = 20% faster)
        adjustedHours:
          type: number
          description: Final estimated hours after experience adjustment
        totalLaborCost:
          type: number
          description: Total labor cost in USD (adjustedHours * hourlyRate)
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/LaborTask"
          description: Task-by-task hour breakdown
        experienceNote:
          type: string
          description: Short note explaining the experience adjustment
      required:
        - baseHours
        - experienceMultiplier
        - adjustedHours
        - totalLaborCost
        - breakdown
        - experienceNote
    MaterialItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          description: Material name
        description:
          type: string
          description: Brief description or specification
        category:
          type: string
          description: Material category (e.g., "Lumber", "Concrete", "Hardware")
        quantity:
          type: number
          description: Amount needed
        unit:
          type: string
          description: Unit of measurement (e.g., "sq ft", "pieces", "bags", "gallons")
        unitPrice:
          type: number
          description: Price per unit in USD
        totalPrice:
          type: number
          description: Total price for this line item in USD
        storeName:
          type: string
          description: Suggested retail store (e.g., "Home Depot", "Lowe's")
        notes:
          type: ["string", "null"]
          description: Any additional notes about this material
      required:
        - id
        - name
        - description
        - category
        - quantity
        - unit
        - unitPrice
        - totalPrice
        - storeName
    EstimateResult:
      type: object
      properties:
        jobSummary:
          type: string
          description: AI summary of the job
        materials:
          type: array
          items:
            $ref: "#/components/schemas/MaterialItem"
        grandTotal:
          type: number
          description: Total cost of all materials in USD
        laborEstimate:
          oneOf:
            - $ref: "#/components/schemas/LaborEstimate"
            - type: "null"
          description: Labor estimate if hourly rate was provided
        disclaimer:
          type: string
          description: Pricing disclaimer
      required:
        - jobSummary
        - materials
        - grandTotal
        - disclaimer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message