# lib/api-zod/src/generated/types/estimateRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EstimateRequest {
  /** Description of the construction or renovation job */
  jobDescription: string;
  /**
     * Optional location for regional pricing context (e.g., "Texas", "California")
     * @nullable
     */
  location?: string | null;
  /**
     * Contractor's hourly labor rate in USD
     * @nullable
     */
  hourlyRate?: number | null;
  /**
     * Contractor's total years of experience (can be fractional, e.g., 0.5 for 6 months)
     * @nullable
     */
  yearsExperience?: number | null;
}