# lib/api-zod/src/generated/types/estimateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LaborEstimate } from './laborEstimate';
import type { MaterialItem } from './materialItem';

export interface EstimateResult {
  /** AI summary of the job */
  jobSummary: string;
  materials: MaterialItem[];
  /** Total cost of all materials in USD */
  grandTotal: number;
  /** Labor estimate if hourly rate was provided */
  laborEstimate?: LaborEstimate | null;
  /** Pricing disclaimer */
  disclaimer: string;
}