# lib/api-zod/src/generated/types/laborEstimate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LaborTask } from './laborTask';

export interface LaborEstimate {
  /** Baseline hours for an average contractor */
  baseHours: number;
  /** Speed multiplier based on experience (e.g., 0.8 = 20% faster) */
  experienceMultiplier: number;
  /** Final estimated hours after experience adjustment */
  adjustedHours: number;
  /** Total labor cost in USD (adjustedHours * hourlyRate) */
  totalLaborCost: number;
  /** Task-by-task hour breakdown */
  breakdown: LaborTask[];
  /** Short note explaining the experience adjustment */
  experienceNote: string;
}