# lib/api-zod/src/generated/types/laborTask.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LaborTask {
  /** Name of the labor task */
  task: string;
  /** Estimated hours for this task */
  hours: number;
}